requirements.txt
numpy==1.24.4
Pillow==9.4.0
